refactor: split pr.py into query (pr.py) and action (pr_actions.py) modules (#140)
